+++ src/components/MetaVoxAdmin.vue
@@ -18,6 +18,7 @@
 				<CogIcon v-if="tab.id === 'groupfolders'" :size="16" />
 				<ShieldIcon v-if="tab.id === 'permissions'" :size="16" />
 			<ChartBoxIcon v-if="tab.id === 'statistics'" :size="16" />
+				<BackupRestoreIcon v-if="tab.id === 'backup'" :size="16" />
 				{{ tab.name }}
 			</button>
 		</div>
@@ -45,6 +46,10 @@
 
 				<StatisticsSettings
 					v-if="activeTab === 'statistics'" />
+
+				<BackupRestore
+					v-if="activeTab === 'backup'"
+					@notification="showNotification" />
 			</div>
 		</div>
 	</div>
@@ -57,6 +62,7 @@ import FileIcon from 'vue-material-design-icons/File.vue'
 import CogIcon from 'vue-material-design-icons/Cog.vue'
 import ShieldIcon from 'vue-material-design-icons/Shield.vue'
 import ChartBoxIcon from 'vue-material-design-icons/ChartBox.vue'
+import BackupRestoreIcon from 'vue-material-design-icons/BackupRestore.vue'
 
 // Import our custom components
 import GroupfolderMetadataFields from './GroupfolderMetadataFields.vue'
@@ -64,6 +70,7 @@ import FileMetadataFields from './FileMetadataFields.vue'
 import ManageGroupfolders from './ManageGroupfolders.vue'
 import PermissionsManager from './PermissionsManager.vue'
 import StatisticsSettings from './StatisticsSettings.vue'
+import BackupRestore from './BackupRestore.vue'
 
 export default {
 	name: 'MetaVoxAdmin',
@@ -79,6 +86,8 @@ export default {
 		ManageGroupfolders,
 		PermissionsManager,
 		StatisticsSettings,
+		BackupRestore,
+		BackupRestoreIcon,
 	},
 
 	data() {
@@ -90,6 +99,7 @@ export default {
 				{ id: 'groupfolders', name: this.t('metavox', 'Manage Team folders') },
 				{ id: 'permissions', name: this.t('metavox', 'User Permissions') },
 				{ id: 'statistics', name: this.t('metavox', 'Statistics') },
+				{ id: 'backup', name: this.t('metavox', 'Backup & Restore') },
 			]
 		}
 	},